requirements.txt
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.16.0
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
zipp==3.21.0
